bug: css path error fixed

+++ index.js
@@ -1,9 +1,11 @@
 const express = require('express');
 const db = require("./database/queries");
 const app = express();
+const { body, validationResult } = require("express-validator");
 const path = require("node:path");
 const assetsPath = path.join(__dirname, "public");
-const { body, validationResult } = require("express-validator");
+app.use(express.static(assetsPath));
+require('dotenv').config();
 
 const validateInput = [
     body("name").trim().isLength({ min: 1, max: 15 }).withMessage("The name can needs to have from 0 to 15 characters."),
@@ -24,9 +26,8 @@ const categories = [
 
 app.set("views", path.join(__dirname, "views"));
 app.set("view engine", "ejs");
-
 app.use(express.urlencoded({ extended: true }));
-app.use(express.static(assetsPath));
+
 
 async function renderAllItems(req, res){
     const items = await db.getAllItems();
@@ -52,12 +53,22 @@ async function updateItem(req, res, item){
     await db.updateItem(item);
     res.redirect("/");
 }
+async function renderCategory(req, res, category){
+    const items = await db.getAllItemsCategory(category);
+    console.log(items);
+    res.render("category", {items:items, category: category});
+}
+
 app.get('/',(req, res)=>{
     renderAllItems(req, res);
 });
 app.get('/categories',(req, res)=>{
     res.render("categories", {categories: categories});
 });
+app.get('/categories/:category',(req, res)=>{
+    renderCategory(req, res, req.params.category);
+});
+
 app.get('/new',(req, res)=>{
     res.render("addItem", {categories:categories});
 });
@@ -71,14 +82,21 @@ app.post('/new', validateInput, (req, res)=>{
     addItem(req, res, req.body);
 });
 app.get('/delete/:id',(req, res)=>{
-    deleteItem(req, res, req.params.id);
+    res.render('admin', {id: req.params.id});
+    //deleteItem(req, res, req.params.id);
+});
+app.post('/delete/:id',(req, res)=>{
+    console.log("password: " + req.body.password);
+    if(req.body.password == process.env.ADMIN){
+        deleteItem(req, res, req.params.id);
+    }else{
+        res.redirect("/");
+    }
 });
 app.get('/edit/:id',(req, res)=>{
     getElement(req, res, req.params.id);
 });
 app.post('/edit/:id',(req, res)=>{
-    console.log("LETS FUCKING GOO");
-    console.log(req.body);
     updateItem(req, res, {name: req.body.name, desc: req.body.desc, price: req.body.price, stock: req.body.stock, category: req.body.category, id: req.params.id})
 });
